pyproject.toml
[project]
name = "assetcues-invoice-ai"
version = "0.1.0"
description = "Assetcues Invoice Agentic AI — invoice extraction pipeline"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langchain-openai>=0.2.0",
    "google-genai>=1.0.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "opencv-python-headless>=4.10.0",
    "Pillow>=10.4.0",
    "numpy>=1.26.0",
    "pypdfium2>=4.30.0",
    "pillow-heif>=0.18.0",
    "httpx>=0.27.0",
    "tenacity>=9.0.0",
    "structlog>=24.4.0",
    "aiofiles>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
    "bandit>=1.8.0",
    "pip-audit>=2.7.0",
    "safety>=3.2.0",
]

[tool.setuptools.packages.find]
include = ["far_ai_brain*"]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--strict-markers -ra"

[tool.coverage.run]
source = ["far_ai_brain"]
omit = ["far_ai_brain/__pycache__/*"]

[tool.coverage.report]
fail_under = 50
show_missing = true
exclude_lines = ["pragma: no cover", "if __name__", "if TYPE_CHECKING"]

[tool.bandit]
exclude_dirs = ["tests", "training_data"]
skips = ["B101"]
